Handle orgs with no SLDs as a separate case
The reason this is necessary is that if an org has no SLDs then the
corresponding regex we construct will be empty.  Thus if we continue
and perform the database queries then they will match every
certificate in the database.  This is obviously wrong, so instead we
create the requisite data structures with empty lists and zero counts.