Query against trimmed_subjects field instead of subjects
This allows for a couple of simplifications.  First, we need not use
regexes in our queries because we're already comparing against a list
of subjects trimmed down to SLDs.  Second, we no longer need to handle
the case where there are no SLDs separately.  This is because when we
use the '$in' query operator with an empty list it will match nothing,
as opposed to an empty regex which matches everything.