Compute whether the org is a suborg. Don't rely on display_owner.
@dav3r pointed out that display_owner is not a foolproof litmus test
for whether an org is a suborg or not.  The better way to check is to
determine if the last snapshot's ID appears as a parent for the last
snapshot.  If it does not then the org is a suborg, and otherwise it
is not.

I added this calculation to the Python code, and the result is stored
in the report.json file.  The result is in turn used in the mustache
file to customize the report content based on whether or not the org
is actually a suborg.