Only fail for sp!=reject for organizational domains
Instead of simply checking if sp!= reject, or that result with
is_base_domain==False.  This means that only organizational domains
are penalized for sp!=reject.

The reason for this is that, according to RFC7489, "'sp'" will be
ignored for DMARC records published on subdomains of Organizational
Domains due to the effect of the DMARC policy discovery mechanism."
Therefore we have chosen not to penalize subdomains for "sp!=reject"
when considering whether subdomains satisfy a DMARC subdomain policy
of reject.

See here for the details in the RFC:
https://tools.ietf.org/html/rfc7489#section-6.3)