Make snapshot creation multithreaded
We start with the list of reports to generate, then figure out which 
snapshots need to be generated in order to create those reports.  
Finally, we split up the list of snapshots into sublists and launch a 
snapshot-creation thread to process each sublist.  This is similar to 
how we create reports, though that report code still needs to be cleaned 
up to match how this snapshot code now works.

Note that all snapshots (including snapshots for third-party reports) 
are now created from a common create_snapshot() function, which 
simplifies the code slightly.