Use unicodecsv instead of csv module
unicodecsv is a drop-in replacement for the built-in csv module.  We have switched to unicodecsv because we now have unicode data in some of our CSVs (e.g. findings data coming from our vulnerability scanner) and the built-in csv module will throw a UnicodeEncodeError when it encounters those unicode characters.