Make notifications_to_delete list more accurate
This change removes any orgs that just had a notification PDF generated via an ancestor org.

Previously, if an org didn't receive a notification PDF directly, but the org's notification was included in an ancestor's PDF, that org would still be included in the notifications_to_delete list.  This resulted in misleading log output.