Modify function that calls cyhy-snapshot
We now pull individual org IDs from the global list of snapshots_to_generate and pass them into the create_snapshot function instead of passing in the entire list of orgs.

Each thread checks to see if the global list is empty; if it isn't, it grabs an org ID and generates a snapshot for it.  If the global list is empty, the thread exits.