Increase "waiting for commander to pause" timeout
As the volume of CyHy stakeholders has grown, commander loop cycles have increased in duration, to the point where it can take longer than 30 minutes for the commander to complete a loop and pause.  This change increases the waiting timeout period in the create_snapshots_reports_scorecard.py script from 30 minutes to 90 minutes to compensate for those longer cycles.